+++ applications/fishing-map/src/features/map/Map.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useState, useEffect } from 'react'
+import React, { useCallback, useState, useEffect, useMemo } from 'react'
 import { createPortal } from 'react-dom'
 import { scaleLinear } from 'd3-scale'
 import { useSelector, useDispatch } from 'react-redux'
@@ -32,7 +32,7 @@ import PopupWrapper from './popups/PopupWrapper'
 import useViewport, { useMapBounds } from './map-viewport.hooks'
 import styles from './Map.module.css'
 import { SliceInteractionEvent } from './map.slice'
-import { useMapSourceLoaded } from './map-features.hooks'
+import { useHasSourceLoaded } from './map-features.hooks'
 
 import '@globalfishingwatch/mapbox-gl/dist/mapbox-gl.css'
 
@@ -122,25 +122,32 @@ const MapWrapper = (): React.ReactElement | null => {
   const dataviews = useSelector(selectDataviewInstancesResolved)
   const mapLegends = useMapLegend(style, dataviews, hoveredEvent)
 
-  const legendsTranslated = mapLegends?.map((legend) => {
-    const isSquareKm = (legend.gridArea as number) > 50000
-    let label = legend.unit
-    if (!label) {
-      // TODO review this when environmental layers switchs to heatmapAnimated
-      if (legend.generatorType === GeneratorType.HeatmapAnimated) {
-        const gridArea = isSquareKm ? (legend.gridArea as number) / 1000000 : legend.gridArea
-        const gridAreaFormatted = gridArea
-          ? formatI18nNumber(gridArea, {
-              style: 'unit',
-              unit: isSquareKm ? 'kilometer' : 'meter',
-              unitDisplay: 'short',
-            })
-          : ''
-        label = `${i18n.t('common.hour_plural', 'hours')} / ~${gridAreaFormatted}²`
-      }
-    }
-    return { ...legend, label }
-  })
+  const legendsTranslated = useMemo(
+    () =>
+      mapLegends?.map((legend) => {
+        const isSquareKm = (legend.gridArea as number) > 50000
+        let label = legend.unit
+        if (!label) {
+          if (legend.generatorType === GeneratorType.HeatmapAnimated) {
+            const gridArea = isSquareKm ? (legend.gridArea as number) / 1000000 : legend.gridArea
+            const gridAreaFormatted = gridArea
+              ? formatI18nNumber(gridArea, {
+                  style: 'unit',
+                  unit: isSquareKm ? 'kilometer' : 'meter',
+                  unitDisplay: 'short',
+                })
+              : ''
+            if (!legend.unit || legend.unit === 'hours') {
+              label = `${i18n.t('common.hour_plural', 'hours')} / ${gridAreaFormatted}²`
+            } else {
+              label = `${legend.unit} - ${gridAreaFormatted}² per cell`
+            }
+          }
+        }
+        return { ...legend, label }
+      }),
+    [mapLegends]
+  )
 
   const debugOptions = useSelector(selectDebugOptions)
 
@@ -151,7 +158,7 @@ const MapWrapper = (): React.ReactElement | null => {
   // TODO handle also in case of error
   // https://docs.mapbox.com/mapbox-gl-js/api/map/#map.event:sourcedataloading
   const tilesLoading = useTilesLoading(map)
-  const encounterSourceLoaded = useMapSourceLoaded(ENCOUNTER_EVENTS_SOURCE_ID)
+  const encounterSourceLoaded = useHasSourceLoaded(ENCOUNTER_EVENTS_SOURCE_ID)
 
   const getCursor = useCallback(
     (state) => {
@@ -185,7 +192,6 @@ const MapWrapper = (): React.ReactElement | null => {
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [map, debugOptions])
-
   return (
     <div className={styles.container}>
       {<MapScreenshot map={map} />}
@@ -229,7 +235,7 @@ const MapWrapper = (): React.ReactElement | null => {
         </InteractiveMap>
       )}
       <MapControls onMouseEnter={resetHoverState} mapLoading={tilesLoading} />
-      {legendsTranslated?.map((legend) => {
+      {legendsTranslated?.map((legend: any) => {
         const legendDomElement = document.getElementById(legend.id as string)
         if (legendDomElement) {
           return createPortal(