+++ applications/temporalgrid-demo/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @globalfishingwatchapp/temporalgrid-demo
 
+## 0.7.58
+
+### Patch Changes
+
+- Updated dependencies [358004cb]
+  - @globalfishingwatch/layer-composer@5.0.0
+  - @globalfishingwatch/react-hooks@6.0.0
+
+## 0.7.57
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/layer-composer@4.19.2
+  - @globalfishingwatch/react-hooks@5.3.2
+
 ## 0.7.56
 
 ### Patch Changes