+++ packages/layer-composer/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @globalfishingwatch/layer-composer
 
+## 5.0.0
+
+### Major Changes
+
+- 358004cb: Use client-side aggregation with env layers
+
+### Patch Changes
+
+- Updated dependencies [358004cb]
+  - @globalfishingwatch/fourwings-aggregate@3.0.0
+
+## 4.19.2
+
+### Patch Changes
+
+- call cluster tiles api with the new "maxClusterZoom" parameter
+
 ## 4.19.1
 
 ### Patch Changes